pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crypto-bot"
version = "1.0.0"
description = "Advanced cryptocurrency trading bot for Coinbase"
authors = [{name = "Durga Ganesh", email = "dganesh.iit@gmail.com"}]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "coinbase-advanced-py>=1.8.2",
    "pandas>=2.0.0",
    "numpy>=1.20.0",
    "click>=8.0.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "python-dotenv>=0.19.0",
    "rich>=13.0.0",
    "urllib3<2.0"
]

[project.scripts]
crypto-bot = "src.cli.commands:main"

[project.urls]
Homepage = "https://github.com/durga-ganesh/coinbase-bot.git"
Repository = "https://github.com/durga-ganesh/coinbase-bot.git"
Issues = "https://github.com/durga-ganesh/coinbase-bot/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-dir]
"" = "."
